added jpa

+++ user/src/main/java/com/userService/user/service/UserService.java
@@ -8,6 +8,6 @@ import com.userService.user.model.User;
 @Service
 public interface UserService {
 
-	CommonResponse<Integer> addUpdateUser(User user);
+	CommonResponse<Long> addUpdateUser(User user);
 
 }
